# services/gateway/swagger.yaml
openapi: 3.0.0
info:
  title: AI Code Reviewer API
  description: |
    API complète pour l'analyse automatique de code source via Intelligence Artificielle.
    
    **Architecture**: Microservices avec API Gateway
    
    **Services**:
    - Auth Service (Port 3001) - Authentification JWT
    - Code Service (Port 3002) - Gestion des analyses
    - AI Service (Port 3003) - Analyse OpenAI
    - Payment Service (Port 3004) - Paiements Stripe
    
    **Authentication**: Utiliser le header `Authorization: Bearer <token>`
  version: 1.0.0
  contact:
    name: Support AI Code Reviewer
    email: support@aicodereviewer.com
servers:
  - url: http://localhost:3000/api
    description: Développement local
  - url: https://api.aicodereviewer.com/api
    description: Production
...
paths:
...
  /code/submit:
    post:
      summary: Soumettre du code pour analyse
      tags: [Analyse de Code]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code, language]
              properties:
                code:
                  type: string
                  minLength: 10
                  maxLength: 50000
                  example: "function add(a, b) { return a + b; }"
                language:
                  type: string
                  enum: [javascript, typescript, python, java, go, rust, php, cpp, csharp, ruby, swift, kotlin, sql, html, css, json, yaml, xml, shell, other]
                  example: "javascript"
                fileName:
                  type: string
                  example: "utils.js"
                context:
                  type: string
                  description: Contexte optionnel pour l'analyse
      responses:
        202:
          description: Analyse acceptée et en cours de traitement
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  analysisId:
                    type: string
                    format: uuid
                  status:
                    type: string
                    enum: [pending, processing]
                  estimatedTime:
                    type: string
        403:
          description: Crédits insuffisants
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /code/history:
    get:
      summary: Récupérer l'historique des analyses
      tags: [Analyse de Code]
      security:
        - BearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, processing, completed, failed]
        - name: language
          in: query
          schema:
            type: string
      responses:
        200:
          description: Liste des analyses
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      analyses:
                        type: array
                        items:
                          $ref: '#/components/schemas/AnalysisSummary'
                      pagination:
                        $ref: '#/components/schemas/Pagination'
  /code/analysis/{id}:
    get:
      summary: Récupérer le détail d'une analyse
      tags: [Analyse de Code]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: Détail de l'analyse
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/AnalysisDetail'
        404:
          description: Analyse non trouvée

    delete:
      summary: Supprimer une analyse
      tags: [Analyse de Code]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: Analyse supprimée
  /code/stats:
    get:
      summary: Récupérer les statistiques de l'utilisateur
      tags: [Analyse de Code]
      security:
        - BearerAuth: []
      responses:
        200:
          description: Statistiques
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserStats'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtenu lors de la connexion
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
    AnalysisSummary:
      type: object
      properties:
        id:
          type: string
        language:
          type: string
        fileName:
          type: string
        status:
          type: string
          enum: [pending, processing, completed, failed]
        score:
          type: number
          minimum: 0
          maximum: 10
        errorsCount:
          type: integer
        warningsCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        executionTime:
          type: number
    AnalysisDetail:
      type: object
      properties:
        _id:
          type: string
        userId:
          type: string
        code:
          type: string
        language:
          type: string
        fileName:
          type: string
        status:
          type: string
        result:
          $ref: '#/components/schemas/AnalysisResult'
        executionTime:
          type: number
        tokensUsed:
          type: object
          properties:
            input:
              type: integer
            output:
              type: integer
            total:
              type: integer
        creditsUsed:
          type: integer
        createdAt:
          type: string
          format: date-time
    AnalysisResult:
      type: object
      properties:
        score:
          type: number
          minimum: 0
          maximum: 10
        scoreBreakdown:
          type: object
          properties:
            quality:
              type: number
            security:
              type: number
            performance:
              type: number
            readability:
              type: number
        errors:
          type: array
          items:
            $ref: '#/components/schemas/CodeIssue'
        warnings:
          type: array
          items:
            $ref: '#/components/schemas/CodeIssue'
        suggestions:
          type: array
          items:
            $ref: '#/components/schemas/CodeSuggestion'
        securityIssues:
          type: array
          items:
            $ref: '#/components/schemas/SecurityIssue'
        summary:
          type: string
    CodeIssue:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [SYNTAX_ERROR, LOGIC_ERROR, SECURITY, PERFORMANCE, STYLE, BEST_PRACTICE]
        severity:
          type: string
          enum: [CRITICAL, HIGH, MEDIUM, LOW]
        line:
          type: integer
        column:
          type: integer
        description:
          type: string
        fix:
          type: string
    CodeSuggestion:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [OPTIMIZATION, READABILITY, MODERNIZATION, REFACTORING, DOCUMENTATION]
        description:
          type: string
        impact:
          type: string
        beforeCode:
          type: string
        afterCode:
          type: string
        explanation:
          type: string
    SecurityIssue:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        severity:
          type: string
          enum: [CRITICAL, HIGH, MEDIUM, LOW]
        description:
          type: string
        line:
          type: integer
        recommendation:
          type: string
    Pagination:
      type: object
      properties:
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
        hasMore:
          type: boolean
    UserStats:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            totalAnalyses:
              type: integer
            averageScore:
              type: number
            totalErrors:
              type: integer
            totalWarnings:
              type: integer
            languages:
              type: array
              items:
                type: string
            languageDistribution:
              type: array
              items:
                type: object
                properties:
                  language:
                    type: string
                  count:
                    type: integer
            recentAnalyses:
              type: integer
            creditsRemaining:
              type: integer
            isPremium:
              type: boolean